requirements.txt
altair==4.1.0
argon2-cffi @ file:///C:/ci/argon2-cffi_1613038019788/work
astor==0.8.1
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
base58==2.1.0
bleach @ file:///tmp/build/80754af9/bleach_1612211392645/work
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1625831764494/work
chardet==4.0.0
click==7.1.2
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler==0.10.0
decorator==4.4.2
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
entrypoints==0.3
gitdb==4.0.7
GitPython==3.1.18
idna==2.10
imageio==2.9.0
importlib-metadata @ file:///C:/ci/importlib-metadata_1617877486026/work
ipykernel @ file:///C:/ci/ipykernel_1596208728219/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1617121109687/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.6.3
jedi==0.17.0
Jinja2 @ file:///tmp/build/80754af9/jinja2_1624781299557/work
joblib==1.0.1
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213516947/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe @ file:///C:/ci/markupsafe_1621528383308/work
matplotlib==3.4.2
mistune @ file:///C:/ci/mistune_1594373272338/work
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1624479160228/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
networkx==2.5.1
notebook @ file:///C:/ci/notebook_1621528807363/work
numpy==1.21.0
packaging @ file:///tmp/build/80754af9/packaging_1625611678980/work
pandas==1.3.0
pandocfilters @ file:///C:/ci/pandocfilters_1605102427207/work
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.3.1
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1623189609245/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
protobuf==3.17.3
pyarrow==4.0.1
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydeck==0.6.2
Pygments @ file:///tmp/build/80754af9/pygments_1621606182707/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyrsistent @ file:///C:/ci/pyrsistent_1600123688363/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz==2021.1
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
pyzmq==20.0.0
requests==2.25.1
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.0
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==4.0.0
streamlit==0.84.0
terminado==0.9.4
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
threadpoolctl==2.2.0
tifffile==2021.7.2
toml==0.10.2
toolz==0.11.1
tornado @ file:///C:/ci/tornado_1606935947090/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1624965014186/work
tzlocal==2.1
urllib3==1.26.6
validators==0.18.2
watchdog==2.1.3
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
widgetsnbextension==3.5.1
wincertstore==0.2
zipp @ file:///tmp/build/80754af9/zipp_1625570634446/work
